PLUGINS/UCX: Split thread and threadpool engines into separate files
