+++ src/plugins/ucx/meson.build
@@ -22,6 +22,8 @@ ucx_backend_sources = ['config.cpp',
                        'mem_list.cpp',
                        'rkey.cpp',
                        'ucx_backend.cpp',
+                       'ucx_engine_thread.cpp',
+                       'ucx_engine_threadpool.cpp',
                        'ucx_sgl.cpp',
                        'ucx_enums.cpp',
                        'ucx_plugin.cpp',
