+++ src/plugins/ucx/ucx_engine_threadpool.cpp
@@ -1,0 +1,370 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "ucx_engine_threadpool.h"
+#include "ucx_request.h"
+
+#include "common/backend.h"
+#include "common/nixl_log.h"
+
+#include <algorithm>
+#include <atomic>
+#include <future>
+#include <memory>
+#include <ostream>
+#include <string>
+#include <vector>
+#include <asio.hpp>
+
+#include "ucx_utils.h"
+
+struct nixlUcxBackendSharedState;
+
+/*
+ * This class represents a chunk of a composite request.
+ * It is used to encapsulate a batch of requests (subset of the larger batch)
+ * performed by a dedicated worker thread of threadpool. It holds a shared state
+ * with the main request to track its completion status and control the lifetime.
+ */
+class nixlUcxChunkBackendReqH : public nixlUcxBackendReqH {
+public:
+    nixlUcxChunkBackendReqH() : nixlUcxBackendReqH(nullptr) {}
+
+    void
+    startXfer(const std::shared_ptr<nixlUcxBackendSharedState> &shared_state,
+              nixlUcxWorker *worker) {
+        NIXL_ASSERT(sharedState_.get() == nullptr);
+        sharedState_ = shared_state;
+        setWorker(worker);
+    }
+
+    void
+    complete(nixl_status_t status);
+
+    [[nodiscard]] nixl_status_t
+    status() override;
+
+    friend std::ostream &
+    operator<<(std::ostream &os, const nixlUcxChunkBackendReqH &chunk) {
+        std::string id = chunk.getWorker() ? std::to_string(chunk.getWorkerId()) : "none";
+        os << "chunk " << &chunk << "{worker_id: " << id << ", state: " << chunk.sharedState_.get()
+           << "}";
+        return os;
+    }
+
+private:
+    std::shared_ptr<nixlUcxBackendSharedState> sharedState_;
+};
+
+/*
+ * This class represents a shared state between a main request and all of its
+ * chunks. It is used to track the completion status of the request and the
+ * number of pending requests, and to control the lifetime of the chunks.
+ */
+struct nixlUcxBackendSharedState {
+    std::atomic<nixl_status_t> status;
+    std::atomic<size_t> pendingReqs;
+    std::vector<nixlUcxChunkBackendReqH> chunks;
+
+    nixlUcxBackendSharedState() : status(NIXL_SUCCESS), pendingReqs(0) {}
+
+    friend std::ostream &
+    operator<<(std::ostream &os, const nixlUcxBackendSharedState &state) {
+        return os << "state " << &state << "{status: " << state.status.load()
+                  << ", pending=" << state.pendingReqs.load() << "}";
+    }
+};
+
+void
+nixlUcxChunkBackendReqH::complete(const nixl_status_t status) {
+    NIXL_ASSERT(sharedState_.get() != nullptr);
+    if (status != NIXL_SUCCESS) {
+        nixlUcxBackendReqH::release();
+        sharedState_->status.store(status);
+    }
+    sharedState_->pendingReqs.fetch_sub(1);
+    NIXL_TRACE << *this << " completed with status: " << status << ", " << *sharedState_;
+    setWorker(nullptr);
+    sharedState_.reset();
+}
+
+nixl_status_t
+nixlUcxChunkBackendReqH::status() {
+    // First check if entire request was cancelled or failed
+    const nixl_status_t status = sharedState_->status.load();
+    if (status != NIXL_SUCCESS) {
+        return status;
+    }
+    return nixlUcxBackendReqH::status();
+}
+
+/*
+ * This class represents a composite request handle for a UCX backend.
+ * It is used to encapsulate multiple parallel requests performed by dedicated
+ * worker threads of threadpool, with a single request handle, that it returned
+ * to the user.
+ */
+class nixlUcxCompositeBackendReqH : public nixlUcxBackendReqH {
+public:
+    nixlUcxCompositeBackendReqH(nixlUcxWorker *worker, size_t chunk_size, size_t num_chunks)
+        : nixlUcxBackendReqH(worker),
+          sharedState_(std::make_shared<nixlUcxBackendSharedState>()),
+          chunkSize_(chunk_size) {
+        sharedState_->chunks.resize(num_chunks);
+    }
+
+    [[nodiscard]] size_t
+    getChunkSize() const noexcept {
+        return chunkSize_;
+    }
+
+    [[nodiscard]] size_t
+    getNumChunks() const noexcept {
+        return sharedState_ ? sharedState_->chunks.size() : 0;
+    }
+
+    void
+    startXfer() {
+        NIXL_ASSERT(sharedState_->pendingReqs.load() == 0);
+        sharedState_->status.store(NIXL_SUCCESS);
+        sharedState_->pendingReqs.store(getNumChunks());
+    }
+
+    [[nodiscard]] nixlUcxChunkBackendReqH *
+    startChunk(size_t idx, nixlUcxWorker *worker) {
+        nixlUcxChunkBackendReqH *chunk = &sharedState_->chunks[idx];
+        chunk->startXfer(sharedState_, worker);
+        return chunk;
+    }
+
+    [[nodiscard]] bool
+    isComposite() const noexcept override {
+        return true;
+    }
+
+    void
+    release() override {
+        NIXL_TRACE << *this << " releasing";
+        nixlUcxBackendReqH::release();
+        if (sharedState_) {
+            // Set failed status to stop progress chunks
+            sharedState_->status.store(NIXL_ERR_NOT_FOUND);
+            // Reset shared state - it will be effectively released when the last chunk
+            // resets the shared state pointer
+            sharedState_.reset();
+        }
+    }
+
+    [[nodiscard]] nixl_status_t
+    status() override {
+        getWorker()->progressLoop();
+
+        if (sharedState_->pendingReqs.load()) {
+            return NIXL_IN_PROG;
+        }
+
+        const nixl_status_t status = nixlUcxBackendReqH::status();
+        if (status != NIXL_SUCCESS) {
+            return status;
+        }
+
+        return sharedState_->status.load();
+    }
+
+    friend std::ostream &
+    operator<<(std::ostream &os, const nixlUcxCompositeBackendReqH &handle) {
+        os << "composite handle " << &handle << "{chunks: " << handle.getNumChunks();
+        if (handle.sharedState_) {
+            os << ", " << *handle.sharedState_;
+        } else {
+            os << ", state: nullptr";
+        }
+        return os << "}}";
+    }
+
+private:
+    std::shared_ptr<nixlUcxBackendSharedState> sharedState_;
+    size_t chunkSize_;
+};
+
+class nixlUcxDedicatedThread : public nixlUcxThread {
+public:
+    nixlUcxDedicatedThread(nixlUcxEngine *engine, asio::io_context &io)
+        : nixlUcxThread(engine, 1),
+          io_(io) {}
+
+    static nixlUcxDedicatedThread *
+    getDedicatedThread() {
+        return (nixlUcxDedicatedThread *)tlsThread();
+    }
+
+    void
+    addRequest(nixlUcxChunkBackendReqH *handle) {
+        requests_.push_back(handle);
+    }
+
+protected:
+    void
+    run() override {
+        const auto guard = asio::make_work_guard(io_);
+        NIXL_DEBUG << "dedicated " << *this << " running";
+
+        while (!io_.stopped()) {
+            if (!requests_.empty()) {
+                io_.poll_one();
+            } else {
+                NIXL_TRACE << "dedicated " << *this << " waiting for requests";
+                io_.run_one();
+            }
+
+            if (requests_.empty()) {
+                continue;
+            }
+
+            for (auto it = requests_.begin(); it != requests_.end();) {
+                nixl_status_t status = (*it)->status();
+                if (status != NIXL_IN_PROG) {
+                    NIXL_TRACE << "dedicated " << *this << " completing " << *(*it)
+                               << " with status: " << status;
+                    (*it)->complete(status);
+                    it = requests_.erase(it);
+                } else {
+                    ++it;
+                }
+            }
+        }
+
+        if (!requests_.empty()) {
+            NIXL_WARN << "dedicated " << *this << " dropping " << requests_.size()
+                      << " requests on exit";
+            for (auto *req : requests_) {
+                NIXL_INFO << "dropping " << *req;
+                req->complete(NIXL_ERR_BACKEND);
+            }
+            requests_.clear();
+        }
+
+        NIXL_DEBUG << "dedicated " << *this << " exiting";
+    }
+
+private:
+    asio::io_context &io_;
+    std::vector<nixlUcxChunkBackendReqH *> requests_;
+};
+
+nixlUcxThreadPoolEngine::nixlUcxThreadPoolEngine(const nixlBackendInitParams &init_params,
+                                                 size_t num_threads)
+    : nixlUcxThreadEngine(init_params, num_threads) {
+    splitBatchSize_ =
+        nixl::getBackendParamDefaulted(init_params.customParams, "split_batch_size", 1024u);
+
+    const auto dedicated_workers = getDedicatedWorkers();
+    io_.reset(new asio::io_context());
+    dedicatedThreads_.reserve(dedicated_workers.size());
+    for (size_t i = 0; i < dedicated_workers.size(); ++i) {
+        dedicatedThreads_.emplace_back(std::make_unique<nixlUcxDedicatedThread>(this, *io_));
+        dedicatedThreads_.back()->addWorker(dedicated_workers[i].get());
+        dedicatedThreads_.back()->start();
+    }
+}
+
+nixlUcxThreadPoolEngine::~nixlUcxThreadPoolEngine() {
+    if (io_) {
+        io_->stop();
+        for (auto &thread : dedicatedThreads_) {
+            thread->join();
+        }
+    }
+}
+
+nixl_status_t
+nixlUcxThreadPoolEngine::prepXfer(const nixl_xfer_op_t &operation,
+                                  const nixl_meta_dlist_t &local,
+                                  const nixl_meta_dlist_t &remote,
+                                  const std::string &remote_agent,
+                                  nixlBackendReqH *&handle,
+                                  const nixl_opt_b_args_t *opt_args) const {
+    size_t batch_size = local.descCount();
+    if (batch_size < splitBatchSize_) {
+        return nixlUcxEngine::prepXfer(operation, local, remote, remote_agent, handle, opt_args);
+    }
+
+    size_t chunk_size = std::max(batch_size / dedicatedThreads_.size(), splitBatchSize_);
+    size_t num_chunks = (batch_size + chunk_size - 1) / chunk_size;
+
+    const auto comp_handle = new nixlUcxCompositeBackendReqH(
+        getSharedWorker(getSharedWorkerId()).get(), chunk_size, num_chunks);
+    NIXL_TRACE << "created " << *comp_handle;
+    handle = comp_handle;
+    return NIXL_SUCCESS;
+}
+
+nixl_status_t
+nixlUcxThreadPoolEngine::sendXferRange(const nixl_xfer_op_t &operation,
+                                       const nixl_meta_dlist_t &local,
+                                       const nixl_meta_dlist_t &remote,
+                                       const std::string &remote_agent,
+                                       nixlBackendReqH *handle,
+                                       size_t start_idx,
+                                       size_t end_idx) const {
+    const auto int_handle = static_cast<nixlUcxBackendReqH *>(handle);
+    if (!int_handle->isComposite()) {
+        return nixlUcxEngine::sendXferRange(
+            operation, local, remote, remote_agent, handle, start_idx, end_idx);
+    }
+
+    const auto comp_handle = static_cast<nixlUcxCompositeBackendReqH *>(int_handle);
+    comp_handle->startXfer();
+    size_t chunk_size = comp_handle->getChunkSize();
+    NIXL_TRACE << "sending " << *comp_handle;
+
+    std::promise<void> promise;
+    std::future<void> future = promise.get_future();
+    std::atomic<size_t> remaining{comp_handle->getNumChunks()};
+    std::atomic<nixl_status_t> status{NIXL_SUCCESS};
+
+    for (size_t i = 0; i < comp_handle->getNumChunks(); i++) {
+        io_->post([&, i]() {
+            nixlUcxDedicatedThread *thread = nixlUcxDedicatedThread::getDedicatedThread();
+            NIXL_ASSERT(thread != nullptr);
+
+            nixlUcxChunkBackendReqH *chunk_handle =
+                comp_handle->startChunk(i, thread->getWorkers()[0]);
+            NIXL_TRACE << "dedicated " << *thread << " starting " << *chunk_handle;
+
+            size_t start_idx = i * chunk_size;
+            size_t end_idx = std::min(start_idx + chunk_size, (size_t)local.descCount());
+            nixl_status_t ret = nixlUcxEngine::sendXferRange(
+                operation, local, remote, remote_agent, chunk_handle, start_idx, end_idx);
+            if (ret != NIXL_SUCCESS) {
+                status.store(ret);
+                chunk_handle->complete(ret);
+            } else {
+                NIXL_TRACE << "dedicated " << *thread << " sent " << *chunk_handle;
+                thread->addRequest(chunk_handle);
+            }
+
+            if (remaining.fetch_sub(1) == 1) {
+                promise.set_value();
+            }
+        });
+    }
+
+    future.wait();
+    NIXL_TRACE << "sent " << *comp_handle << " with status: " << status.load();
+    return status.load();
+}
